python/requirements.txt
### PYTHON package requirements
opencv-python #includes 'numpy'
torch
segment-anything
#used?
requests
Python-IO
torchvision
Pillow